追加修正

+++ src/main/java/jp/co/ricoh/cotos/commonlib/entity/accounting/Accounting.java
@@ -449,7 +449,7 @@ public class Accounting extends EntityBase {
 	/** 売上数量 */
 	@Max(9999999999L)
 	@ApiModelProperty(value = "売上数量", required = false, position = 85, allowableValues = "range[0,99999999999]")
-	private Integer ffmUserSalesCnt;
+	private Long ffmUserSalesCnt;
 
 	/** ユーザ売上単価 */
 	@DecimalMax("9999999999999999999.99")
@@ -489,7 +489,7 @@ public class Accounting extends EntityBase {
 	/** RJ売上数量 */
 	@Max(9999999999L)
 	@ApiModelProperty(value = "RJ売上数量", required = false, position = 93, allowableValues = "range[0,99999999999]")
-	private Integer ffmRjSalesCnt;
+	private Long ffmRjSalesCnt;
 
 	/** RJ売上単価 */
 	@DecimalMax("9999999999999999999.99")
@@ -529,7 +529,7 @@ public class Accounting extends EntityBase {
 	/** RJ仕入数量 */
 	@Max(9999999999L)
 	@ApiModelProperty(value = "RJ仕入数量", required = false, position = 101, allowableValues = "range[0,99999999999]")
-	private Integer ffmRjPurchaseCnt;
+	private Long ffmRjPurchaseCnt;
 
 	/** RJ仕入単価 */
 	@DecimalMax("9999999999999999999.99")
@@ -569,7 +569,7 @@ public class Accounting extends EntityBase {
 	/** 販売店売上数量 */
 	@Max(9999999999L)
 	@ApiModelProperty(value = "販売店売上数量", required = false, position = 109, allowableValues = "range[0,99999999999]")
-	private Integer ffmShopSalesCnt;
+	private Long ffmShopSalesCnt;
 
 	/** 販売店売上単価 */
 	@DecimalMax("9999999999999999999.99")
@@ -610,7 +610,7 @@ public class Accounting extends EntityBase {
 	@Column(name = "ffm_r_cost_cnt")
 	@Max(9999999999L)
 	@ApiModelProperty(value = "R原価数量", required = false, position = 117, allowableValues = "range[0,99999999999]")
-	private Integer ffmRCostCnt;
+	private Long ffmRCostCnt;
 
 	/** R原価単価 */
 	@Column(name = "ffm_r_cost_price")
@@ -657,7 +657,7 @@ public class Accounting extends EntityBase {
 	/** 手数料数量 */
 	@Max(99999999999L)
 	@ApiModelProperty(value = "手数料数量", required = false, position = 125, allowableValues = "range[0,99999999999]")
-	private Integer ffmCommissionCnt;
+	private Long ffmCommissionCnt;
 
 	/** 手数料単価 */
 	@DecimalMax("9999999999999999999.99")
@@ -1039,7 +1039,7 @@ public class Accounting extends EntityBase {
 	/** 数量 */
 	@Max(99999999999L)
 	@ApiModelProperty(value = "数量", required = false, position = 199, allowableValues = "range[0,99999999999]")
-	private Integer cubicCount;
+	private Long cubicCount;
 
 	/** 取引金額 */
 	@DecimalMax("9999999999999999999.99")
